extract token inputs to separate component
